refactor `optimize` to `temporary`
