+++ assets/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: edu.cornell.gdiac.optimize.desktop.DesktopLauncher
+Main-Class: edu.cornell.gdiac.temporary.desktop.DesktopLauncher
