+++ core/src/edu/cornell/gdiac/temporary/GameObject.java
@@ -21,7 +21,7 @@
  * Based on original Optimization Lab by Don Holden, 2007
  * LibGDX version, 2/2/2015
  */
-package edu.cornell.gdiac.optimize;
+package edu.cornell.gdiac.temporary;
 
 import com.badlogic.gdx.math.*;
 import com.badlogic.gdx.graphics.*;
