+++ core/src/edu/cornell/gdiac/temporary/RandomController.java
@@ -13,7 +13,7 @@
  * Based on original Optimization Lab by Don Holden, 2007
  * LibGDX version, 2/2/2015
  */
-package edu.cornell.gdiac.optimize;
+package edu.cornell.gdiac.temporary;
 
 import java.util.Random;
