+++ core/src/edu/cornell/gdiac/temporary/entity/Note.java
@@ -16,9 +16,9 @@
  * Based on original Optimization Lab by Don Holden, 2007
  * LibGDX version, 2/2/2015
  */
-package edu.cornell.gdiac.optimize.entity;
+package edu.cornell.gdiac.temporary.entity;
 
-import edu.cornell.gdiac.optimize.*;
+import edu.cornell.gdiac.temporary.*;
 import edu.cornell.gdiac.util.*;
 
 import com.badlogic.gdx.math.*;
